Few fixes on const parameters and return value
Former-commit-id: 3b9275bb8a48b2fbd40a3e278c51d95219af11c4

+++ roborobo3/prj/Negociate/src/NegociateWorldObserver.cpp
@@ -240,7 +240,7 @@ void NegociateWorldObserver::stepEvolution()
 std::vector<std::pair<int, double>> NegociateWorldObserver::getSortedFitnesses() const
 {
     std::vector<std::pair<int, double>> fitnesses(m_individuals.size());
-    for (int i = 0; i < m_individuals.size(); i++)
+    for (size_t i = 0; i < m_individuals.size(); i++)
     {
         fitnesses[i].first = i;
         fitnesses[i].second = m_fitnesses[i];
